requirements.txt
faiss-cpu
sentence-transformers
ollama